requirements.txt
fastapi[standard]==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
python-logging-loki==0.3.1
requests==2.31.0